[WEB-08] routed to contact page when button is clicked
