+++ src/app/(sites)/contact/page.tsx
@@ -1,9 +1,18 @@
-import { useRouter } from "next/navigation";
+import Link from "next/link";
 import React from "react";
 
 const Contact = () => {
-  const router = useRouter();
-  return <div> Contact page</div>;
+  return (
+    <div>
+      <h1 className="text-5xl ">Contact </h1>
+      <Link
+        href="
+    /"
+      >
+        HOME
+      </Link>
+    </div>
+  );
 };
 
 export default Contact;
