+++ src/components/Navbar/Navbar.tsx
@@ -1,8 +1,7 @@
 import Image from "next/image";
 import Link from "next/link";
 import React from "react";
-import { SideMenu, Platforms, Resources, Solutions } from ".";
-import { PrimaryButton } from "../Reusable";
+import { SideMenu, Platforms, Resources, Solutions, ContactUs } from ".";
 
 const Navbar = () => {
   return (
@@ -31,11 +30,7 @@ const Navbar = () => {
           </div>
 
           <div className="flex-center gap-3">
-            <PrimaryButton
-              type="button"
-              title="CONTACT US"
-              className="px-5 py-2 lg:px-7 lg:py-2.5 text-xs lg:text-lg"
-            />
+            <ContactUs />
             <SideMenu />
           </div>
         </div>
